refactor: extract TranscriptQualityFindingsView.swift from TranscriptRecoveryAndHistoryViews.swift
Extracts the TranscriptQualityFindingsView SwiftUI view (29 lines) from
TranscriptRecoveryAndHistoryViews.swift into its own file.

Byte-identical move; SHA verified against origin/main.

Reduces TranscriptRecoveryAndHistoryViews.swift from 144 to 112 lines.

Tests: 667 (unchanged).

⚠️ Manual smoke check recommended before merge: verify quality findings
display renders identically when a session has quality warnings.
